refactor(default-theme): Extract PageNav component.
